+++ lib/features/dashboard/presentation/page/dashboard_page.dart
@@ -1,7 +1,12 @@
 import 'package:fl_chart/fl_chart.dart';
-import 'package:fortuno/core/core.dart';
-import 'package:fortuno/core/widgets/custom_tab.dart';
-import 'package:fortuno/features/auth/presentations/bloc/auth_bloc.dart';
+
+import '../../../../core/core.dart';
+import '../../../../core/widgets/custom_tab.dart';
+import '../../../auth/presentations/bloc/auth_bloc.dart';
+
+part '../widgets/dashboard_app_bar.dart';
+part '../widgets/most_product_widget.dart';
+part '../widgets/total_order_widget.dart';
 
 class DashboardPage extends StatelessWidget {
   static const path = '/dashboard-page';
@@ -424,249 +429,6 @@ class DashboardPage extends StatelessWidget {
           ),
         ),
       ],
-    );
-  }
-}
-
-class DashboardAppBar extends StatelessWidget {
-  const DashboardAppBar({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: anchorCustomAppBar,
-      width: double.infinity,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              Expanded(
-                child: Text("Dashboard", style: context.textTheme.titleLarge),
-              ),
-              Text(formatDate(DateTime.now())),
-              SizedBox(width: kSizeMS),
-              ElevatedButton(
-                onPressed: () {
-                  context.read<AuthBloc>().add(OnSignOutEvent());
-                },
-                child: Row(
-                  children: [
-                    Icon(CupertinoIcons.square_arrow_right, color: darkColor),
-                    SizedBox(width: 6),
-                    Text(
-                      "Keluar",
-                      style: context.textTheme.bodySmall?.copyWith(
-                        color: darkColor,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-          SizedBox(height: kSizeSS),
-          CustomTab(
-            value: 'Minggu',
-            menus: ["Minggu", "Bulan", "Tahun"],
-            changeMenu: (menu) {},
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class TotalOrderWidget extends StatelessWidget {
-  const TotalOrderWidget({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return CustomCard(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Row(
-                children: [
-                  Container(
-                    padding: EdgeInsets.all(5),
-                    decoration: BoxDecoration(
-                      color: secondaryColor,
-                      borderRadius: BorderRadius.circular(5),
-                    ),
-                    child: Icon(
-                      CupertinoIcons.cart,
-                      size: kSizeM,
-                      color: whiteColor,
-                    ),
-                  ),
-                  SizedBox(width: kSizeMS),
-                  Text(
-                    "Total Orders",
-                    style: context.textTheme.bodySmall?.copyWith(
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                ],
-              ),
-              Row(
-                children: [
-                  Icon(CupertinoIcons.arrow_down_right, color: secondaryColor),
-                  Text(
-                    "2%",
-                    style: context.textTheme.bodySmall?.copyWith(
-                      color: secondaryColor,
-                    ),
-                  ),
-                ],
-              ),
-            ],
-          ),
-          Spacer(),
-          Text("100", style: context.textTheme.displayLarge),
-          Spacer(),
-          // Text(
-          //   "lebih banyak 2 pesanan dari pada bulan lalu",
-          //   style: context.textTheme.bodySmall?.copyWith(
-          //     color: lightGrey2,
-          //     fontStyle: FontStyle.italic,
-          //   ),
-          //   textAlign: TextAlign.end,
-          // ),
-        ],
-      ),
-    );
-  }
-}
-
-class MostProductWidget extends StatelessWidget {
-  const MostProductWidget({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final data = [
-      "Paket Hemat Berkah 16K Ayam Bakar",
-      "Paket Hemat Berkah 16K Ayam Goreng",
-      "Paket Kue 10K",
-    ];
-    return CustomCard(
-      width: double.infinity,
-      margin: [anchorRightContent, anchorBottomContent].merge,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              Container(
-                padding: EdgeInsets.all(5),
-                decoration: BoxDecoration(
-                  color: darkColor,
-                  borderRadius: BorderRadius.circular(5),
-                ),
-                child: Icon(
-                  CupertinoIcons.heart_fill,
-                  size: kSizeM,
-                  color: primaryColor,
-                ),
-              ),
-              SizedBox(width: kSizeMS),
-              Text(
-                "Paling sering dipesan",
-                style: context.textTheme.bodySmall?.copyWith(
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ],
-          ),
-          SizedBox(height: kSizeM),
-          Expanded(
-            child: ListView.builder(
-              padding: EdgeInsets.zero,
-              physics: BouncingScrollPhysics(),
-              itemCount: data.length,
-              itemBuilder: (context, index) {
-                return InkWell(
-                  onTap: () {},
-                  child: Container(
-                    margin: EdgeInsets.only(top: kSizeS),
-                    padding: EdgeInsets.all(kSizeS),
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadiusDirectional.circular(
-                        kDefaultRadius * 0.5,
-                      ),
-                      color: lightGrey8,
-                    ),
-                    child: Row(
-                      crossAxisAlignment: CrossAxisAlignment.center,
-                      children: [
-                        Container(
-                          width: 50,
-                          height: 50,
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadiusDirectional.circular(
-                              kDefaultRadius * 0.5,
-                            ),
-                            image: DecorationImage(
-                              image: NetworkImage(
-                                "https://cdn.prod.website-files.com/6346686a30512816becfbf47/671ed63b8d688b6b1d10a0ec_freepik__retouch__32461.png",
-                              ),
-                            ),
-                          ),
-                        ),
-                        SizedBox(width: kSizeS),
-                        Expanded(
-                          child: Row(
-                            children: [
-                              Expanded(
-                                child: Text(
-                                  data[index],
-                                  style: context.textTheme.bodySmall?.copyWith(
-                                    fontWeight: FontWeight.bold,
-                                    fontSize: kSizeS * 0.9,
-                                  ),
-                                ),
-                              ),
-                              SizedBox(width: kSizeM),
-                              Column(
-                                crossAxisAlignment: CrossAxisAlignment.end,
-                                children: [
-                                  Text(
-                                    "20x",
-                                    style: context.textTheme.bodySmall
-                                        ?.copyWith(
-                                          fontWeight: FontWeight.bold,
-                                          fontSize: kSizeS,
-                                          color: secondaryColor,
-                                        ),
-                                    textAlign: TextAlign.justify,
-                                  ),
-                                  Text(
-                                    "200 pcs",
-                                    style: context.textTheme.bodySmall
-                                        ?.copyWith(
-                                          fontWeight: FontWeight.bold,
-                                          fontSize: kSizeS,
-                                          color: secondaryColor,
-                                        ),
-                                    textAlign: TextAlign.justify,
-                                  ),
-                                ],
-                              ),
-                            ],
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                );
-              },
-            ),
-          ),
-        ],
-      ),
     );
   }
 }
